chore(test): split installer tests
